feat: refactor SQL agent and operational tools into dedicated modules for improved maintainability
